requirements.txt
streamlit 
streamlit-authenticator
PyYAML
pandas
numpy
yfinance
bcrypt
plotly
openpyxl
pyotp
qrcode
